+++ classes/plugin/AdminController.php
@@ -22,6 +22,7 @@ use Grav\Common\Page\Medium\Medium;
 use Grav\Common\Page\Page;
 use Grav\Common\Page\Pages;
 use Grav\Common\Page\Collection;
+use Grav\Common\Plugins;
 use Grav\Common\Security;
 use Grav\Common\User\Interfaces\UserCollectionInterface;
 use Grav\Common\User\Interfaces\UserInterface;
@@ -35,6 +36,7 @@ use PicoFeed\Parser\MalformedXmlException;
 use Psr\Http\Message\ResponseInterface;
 use RocketTheme\Toolbox\Event\Event;
 use RocketTheme\Toolbox\File\File;
+use RocketTheme\Toolbox\File\YamlFile;
 use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 use Twig\Loader\FilesystemLoader;
 
@@ -629,10 +631,8 @@ class AdminController extends AdminBaseController
             return false;
         }
 
-        // Filter value and save it.
-        $this->post = ['enabled' => true];
-        $obj        = $this->prepareData($this->post);
-        $obj->save();
+        $type = $this->getDataType();
+        $this->updatePluginState($type, ['enabled' => true]);
 
         $this->post = ['_redirect' => 'plugins'];
         if ($this->grav['uri']->param('redirect')) {
@@ -662,10 +662,8 @@ class AdminController extends AdminBaseController
             return false;
         }
 
-        // Filter value and save it.
-        $this->post = ['enabled' => false];
-        $obj        = $this->prepareData($this->post);
-        $obj->save();
+        $type = $this->getDataType();
+        $this->updatePluginState($type, ['enabled' => false]);
 
         $this->post = ['_redirect' => 'plugins'];
         $this->admin->setMessage($this->admin::translate('PLUGIN_ADMIN.SUCCESSFULLY_DISABLED_PLUGIN'), 'info');
@@ -673,6 +671,30 @@ class AdminController extends AdminBaseController
         Cache::clearCache('invalidate');
 
         return true;
+    }
+
+    /**
+     * @param string $type
+     * @param array $value
+     * @return void
+     */
+    protected function updatePluginState(string $type, array $value): void
+    {
+        $obj = Plugins::get(preg_replace('|plugins/|', '', $type));
+        if (null === $obj) {
+            throw new \RuntimeException("Plugin '{$type}' doesn't exist!");
+        }
+
+        /** @var UniformResourceLocator $locator */
+        $locator = $this->grav['locator'];
+
+        // Configuration file will be saved to the existing config stream.
+        $filename = $locator->findResource('config://') . "/{$type}.yaml";
+
+        $file = YamlFile::instance($filename);
+        $contents = $value + $file->content();
+
+        $file->save($contents);
     }
 
     /**